+++ src/main/kotlin/com/cmgapps/license/reporter/MarkdownReport.kt
@@ -6,26 +6,46 @@
 
 package com.cmgapps.license.reporter
 
+import com.cmgapps.license.helper.LicensesHelper
+import com.cmgapps.license.helper.getLicenseText
+import com.cmgapps.license.helper.logLicenseWarning
+import com.cmgapps.license.helper.toLicensesMap
 import com.cmgapps.license.model.Library
+import org.gradle.api.logging.Logger
 
-internal class MarkdownReport(libraries: List<Library>) : Report(libraries) {
+internal class MarkdownReport(
+    libraries: List<Library>,
+    private val logger: Logger,
+) : Report(libraries) {
     override fun generate() = buildString {
         append("# Open source licenses\n")
-        append("### Notice for packages:\n")
-        libraries.forEach { library ->
-            append(library.name)
-            append(' ')
-            append('_')
-            append(library.mavenCoordinates.version)
-            append("_:")
-            library.licenses.forEach { license ->
-                append("\n* ")
-                append(license.name)
-                append(" (")
-                append(license.url)
-                append(")")
+        append("## Notice for packages")
+        libraries.toLicensesMap().forEach { (license, libraries) ->
+            libraries.asSequence().sortedBy { it.name }.forEach { library ->
+                append('\n')
+                append("* ")
+                append(library.name ?: library.mavenCoordinates.identifierWithoutVersion)
             }
-            append("\n\n")
+
+            val licenseUrl = license.url
+            val licenseName = license.name
+
+            append("\n```\n")
+            when {
+                LicensesHelper.LICENSE_MAP.containsKey(licenseUrl) -> {
+                    append(LicensesHelper.LICENSE_MAP[licenseUrl].getLicenseText() ?: "")
+                }
+                LicensesHelper.LICENSE_MAP.containsKey(licenseName) -> {
+                    append(LicensesHelper.LICENSE_MAP[licenseName].getLicenseText() ?: "")
+                }
+                else -> {
+                    logger.logLicenseWarning(license, libraries)
+                    append(licenseName)
+                    append("\n")
+                    append(licenseUrl)
+                }
+            }
+            append("\n```\n")
         }
     }
 }